+++ django_goodies/auth.py
@@ -100,4 +100,7 @@ class ObjectPermissionsBackend(object):
         if not obj:
             return False  # not dealing with non-object permissions
         
-        return self._get_object_permission(perm, user_obj, obj, 'user') or self._get_object_permission(perm, user_obj, obj, 'group')
+        return (
+            self._get_object_permission(perm, user_obj, obj, 'user')
+            or self._get_object_permission(perm, user_obj, obj, 'group')
+        )